requirements.txt
Django>=4.2,<6.0
WeasyPrint>=61.0
qrcode>=7.4
Pillow>=10.0
psycopg2-binary>=2.9,<3.0
python-dotenv>=1.0
